MC_Crafting/requirements.txt
Pillow~=8.2.0
colorama~=0.4.4